Kiem tra so du khach hang truoc khi thanh toan va toi uu hoa gio hang

+++ page_cart.php
@@ -8,11 +8,11 @@ cart();
     <div>
         <?php
         $total = 0;
-        $username=$_SESSION['username'];
-        $select_makh="select * from `taikhoan` where tendangnhap='$username'";
-        $select_makh_run=mysqli_query($con,$select_makh);
-        $row_makh=mysqli_fetch_assoc($select_makh_run);
-        $get_makh=$row_makh['MAKH'];
+        $username = $_SESSION['username'];
+        $select_makh = "select * from `taikhoan` where tendangnhap='$username'";
+        $select_makh_run = mysqli_query($con, $select_makh);
+        $row_makh = mysqli_fetch_assoc($select_makh_run);
+        $get_makh = $row_makh['MAKH'];
         $select_cart = "select * from `giohang` where MAKH='$get_makh'";
         $select_cart_run = mysqli_query($con, $select_cart);
         $count_row = mysqli_num_rows($select_cart_run);
@@ -49,52 +49,47 @@ cart();
                     $gia = $row_product['GIA'];
                     $total += $gia;
         ?>
-        <tr class="table_content">
-            <td><input type="checkbox" name="select[]" value="<?php echo $masp ?>"></td>
-            <td>
-                <img style="width: 60%; height: 60%; object-fit: contain;" src="./<?php echo $url_image; ?>"
-                    alt="<?php echo $tensp ?>">
-            </td>
-            <td><?php echo $tensp; ?></td>
-            <td><?php echo $mau; ?></td>
+                    <tr class="table_content">
+                        <td><input type="checkbox" name="select[]" value="<?php echo $masp ?>"></td>
+                        <td>
+                            <img style="width: 60%; height: 60%; object-fit: contain;" src="./<?php echo $url_image; ?>" alt="<?php echo $tensp ?>">
+                        </td>
+                        <td><?php echo $tensp; ?></td>
+                        <td><?php echo $mau; ?></td>
 
-            <td>
-                <div class="cart_quantity">
-                    <?php echo $soluong ?>
-                </div>
-            </td>
-
-
-            <td><?php $dongia=$soluong*$gia; echo currency_format($dongia); ?> đ</td>
+                        <td>
+                            <div class="cart_quantity">
+                                <?php echo $soluong ?>
+                            </div>
+                        </td>
 
 
-            <td>
-                <input type="submit" value="Thanh toán" name="Thanh_toan"
-                    class="bg-info p-2 border-0 my-2 px-2 button btnshoppingcart">
-                <input type="submit" value="Thanh toán" name="Thanh_toan"
-                    class="bg-info btn-sm button_sm p-1 border-0 my-1 px-2">
-                <input type="submit" value="Xóa" name="xoa" class="bg-info p-2 border-0 button btnshoppingcart">
-                <input type="submit" value="Xóa" name="xoa" class="bg-info btn-sm button_sm p-1 border-0">
+                        <td><?php $dongia = $soluong * $gia;
+                            echo currency_format($dongia); ?> đ</td>
 
-            </td>
-        </tr>
-        <?php
+
+                        <td>
+                            <input type="submit" value="Thanh toán" name="Thanh_toan" class="bg-info p-2 border-0 my-2 px-2 button btnshoppingcart">
+                            <input type="submit" value="Thanh toán" name="Thanh_toan" class="bg-info btn-sm button_sm p-1 border-0 my-1 px-2">
+                            <input type="submit" value="Xóa" name="xoa" class="bg-info p-2 border-0 button btnshoppingcart">
+                            <input type="submit" value="Xóa" name="xoa" class="bg-info btn-sm button_sm p-1 border-0">
+
+                        </td>
+                    </tr>
+            <?php
                 }
             }
             ?>
-        </tbody>
-        </table>
-        <div class="d-flex">
-            <h4 class="px-3">Tổng: <strong
-                    class="text-info shoppingcart_h4"><?php echo currency_format($total) . " đ"; ?></strong></h4>
+            </tbody>
+            </table>
+            <div class="d-flex">
+                <h4 class="px-3">Tổng: <strong class="text-info shoppingcart_h4"><?php echo currency_format($total) . " đ"; ?></strong></h4>
 
-            <a href="#"><button type="submit" name="thanhtoantatca"
-                    class="bg-info p-2 border-0 mx-4 btnshoppingcart">Thanh toán tất cả</button></a>
-            <a class="bg-info p-2 border-0 btnshoppingcart"
-                style="width: 100px; text-align: center; text-decoration: none;" href="./index.php">Thoát</a>
+                <a href="#"><button type="submit" name="thanhtoantatca" class="bg-info p-2 border-0 mx-4 btnshoppingcart">Thanh toán tất cả</button></a>
+                <a class="bg-info p-2 border-0 btnshoppingcart" style="width: 100px; text-align: center; text-decoration: none;" href="./index.php">Thoát</a>
 
-        </div>
-        </form>
+            </div>
+            </form>
         <?php
         } else {
             echo "<div class='cart_list cart_list--no-cart'>
@@ -124,8 +119,7 @@ cart();
                         }
                     }
                 } else {
-                    echo "<script>window.open('page_shopping_cart.php','_self') </script>";
-                    echo "kop dc";
+                    echo "<script>alert('Vui lòng chọn sản phẩm cần thao tác!') </script>";
                 }
             }
         }
@@ -133,30 +127,28 @@ cart();
         function check_thongtin()
         {
             global $con;
-            if(isset($_POST['Thanh_toan']) or isset($_POST['thanhtoantatca']))
-            {
-            $username=$_SESSION['username'];
-            $select_makh="select * from `taikhoan` where tendangnhap='$username'";
-            $select_makh_run=mysqli_query($con,$select_makh);
-            $row_taikhoan=mysqli_fetch_assoc($select_makh_run);
-            $get_makh=$row_taikhoan['MAKH'];
-            $select_khachhang="select * from `khachhang` where MAKH='$get_makh'";
-            $select_khachhang_run=mysqli_query($con,$select_khachhang);
-            $row_khachhang=mysqli_fetch_assoc($select_khachhang_run);
-            $get_hoten=$row_khachhang['HOTEN'];
-            $get_dchi=$row_khachhang['DCHI'];
-            $get_sodt=$row_khachhang['SODT'];
-            $get_gioitinh=$row_khachhang['GIOITINH'];
-            $get_socccd=$row_khachhang['SOCCCD'];
-            if($get_hoten=='' or $get_dchi=='' or $get_sodt=='' or $get_gioitinh=='' or $get_socccd=='')
-            {
-                echo "<script> alert('Bạn phải cập nhập đầy đủ thông tin trước khi thanh toán') </script>";
-                echo "<script>window.open('user.php?profile','_self')</script>>";
-                exit(0);
+            if (isset($_POST['Thanh_toan']) or isset($_POST['thanhtoantatca'])) {
+                $username = $_SESSION['username'];
+                $select_makh = "select * from `taikhoan` where tendangnhap='$username'";
+                $select_makh_run = mysqli_query($con, $select_makh);
+                $row_taikhoan = mysqli_fetch_assoc($select_makh_run);
+                $get_makh = $row_taikhoan['MAKH'];
+                $select_khachhang = "select * from `khachhang` where MAKH='$get_makh'";
+                $select_khachhang_run = mysqli_query($con, $select_khachhang);
+                $row_khachhang = mysqli_fetch_assoc($select_khachhang_run);
+                $get_hoten = $row_khachhang['HOTEN'];
+                $get_dchi = $row_khachhang['DCHI'];
+                $get_sodt = $row_khachhang['SODT'];
+                $get_gioitinh = $row_khachhang['GIOITINH'];
+                $get_socccd = $row_khachhang['SOCCCD'];
+                if ($get_hoten == '' or $get_dchi == '' or $get_sodt == '' or $get_gioitinh == '' or $get_socccd == '') {
+                    echo "<script> alert('Bạn phải cập nhập đầy đủ thông tin trước khi thanh toán') </script>";
+                    echo "<script>window.open('user.php?profile','_self')</script>>";
+                    exit(0);
+                }
             }
         }
-        }
-        echo $check_thongtin=check_thongtin();
+        echo $check_thongtin = check_thongtin();
         function thanhtoan()
         {
             global $con;
@@ -173,66 +165,73 @@ cart();
                         $row_sanpham = mysqli_fetch_assoc($select_sanpham_run);
                         $get_gia = $row_sanpham['GIA'];
                         $total += $get_sl * $get_gia;
-
                     }
+
                     $get_username = $_SESSION['username'];
                     // lấy mã khách hàng
                     $select_khachhang = "select * from `taikhoan` where tendangnhap='$get_username'";
                     $select_khachhang_run = mysqli_query($con, $select_khachhang);
                     $row_khachhang = mysqli_fetch_assoc($select_khachhang_run);
                     $get_makh = $row_khachhang['MAKH'];
-                    //chèn hóa đơn
-                    date_default_timezone_set('Asia/Ho_Chi_Minh');
-                    $get_date = date("Y-m-d H:i:s");
-                    $insert_hoadon = "insert into `hoadon` (NGHD,MAKH,TRIGIA,TRANGTHAI) values ('$get_date','$get_makh','$total','-2')";
-                    $insert_hoadon_run = mysqli_query($con, $insert_hoadon);
-                    //lấy số hd mới dc tạo
-                    $select_hoadon = "select MAX(SOHD) from `hoadon`";
-                    $select_hoadon_run = mysqli_query($con, $select_hoadon);
-                    $row_hoadon = mysqli_fetch_assoc($select_hoadon_run);
-                    $get_sohd = $row_hoadon['MAX(SOHD)'];
-                    foreach ($_POST['select'] as $remove_id) {
-                        //lấy số lượng
-                        $select_giohang = "select * from `giohang` where MASP='$remove_id'";
-                        $select_giohang_run = mysqli_query($con, $select_giohang);
-                        $row_giohang = mysqli_fetch_assoc($select_giohang_run);
-                        $get_sl = $row_giohang['SOLUONG'];
-                        //chèn vào cthd
-                        $insert_cthd = "insert into `cthd` (SOHD,MASP,SL) value ('$get_sohd','$remove_id','$get_sl')";
-                        $insert_cthd_run = mysqli_query($con, $insert_cthd);
-                        //chèn xong thì xóa
-                        $delete_query = "delete from `giohang` where MASP='$remove_id'";
-                        $delete_query = mysqli_query($con, $delete_query);
-                        if ($delete_query) {
-                            echo "<script>window.open('page_shopping_cart.php','_self') </script>";
+
+                    //Kiểm tra số dư
+                    $sql_sodu = "select * from `khachhang` where MAKH = '$get_makh'";
+                    $run = mysqli_query($con, $sql_sodu);
+                    $row_sodu = mysqli_fetch_assoc($run);
+                    $get_sodu = $row_sodu['SODU'];
+                    if ($get_sodu > $total) {
+                        //chèn hóa đơn
+                        date_default_timezone_set('Asia/Ho_Chi_Minh');
+                        $get_date = date("Y-m-d H:i:s");
+                        $insert_hoadon = "insert into `hoadon` (NGHD,MAKH,TRIGIA,TRANGTHAI) values ('$get_date','$get_makh','$total','-2')";
+                        $insert_hoadon_run = mysqli_query($con, $insert_hoadon);
+                        //lấy số hd mới dc tạo
+                        $select_hoadon = "select MAX(SOHD) from `hoadon`";
+                        $select_hoadon_run = mysqli_query($con, $select_hoadon);
+                        $row_hoadon = mysqli_fetch_assoc($select_hoadon_run);
+                        $get_sohd = $row_hoadon['MAX(SOHD)'];
+                        foreach ($_POST['select'] as $remove_id) {
+                            //lấy số lượng
+                            $select_giohang = "select * from `giohang` where MASP='$remove_id'";
+                            $select_giohang_run = mysqli_query($con, $select_giohang);
+                            $row_giohang = mysqli_fetch_assoc($select_giohang_run);
+                            $get_sl = $row_giohang['SOLUONG'];
+                            //chèn vào cthd
+                            $insert_cthd = "insert into `cthd` (SOHD,MASP,SL) value ('$get_sohd','$remove_id','$get_sl')";
+                            $insert_cthd_run = mysqli_query($con, $insert_cthd);
+                            //chèn xong thì xóa
+                            $delete_query = "delete from `giohang` where MASP='$remove_id'";
+                            $delete_query = mysqli_query($con, $delete_query);
+                            if ($delete_query) {
+                                echo "<script>window.open('page_shopping_cart.php','_self') </script>";
+                            }
                         }
+                    } else {
+                        echo "<script>alert('Số dư củ của bạn không đủ! Vui lòng nạp thêm tiền để thanh toán!')</script>";
+                        echo "<script>window.open('user.php?payment') </script>";
                     }
                 } else {
-                    echo "<script>window.open('user.php?purchar&type=-3','_self') </script>";
-                    
+                    echo "<script>alert('Vui lòng chọn sản phẩm cần thao tác!') </script>";
                 }
-            
-             }
+            }
         }
         echo $insert = thanhtoan();
         function thanhtoantatca()
         {
             global $con;
-            $total=0;
+            $total = 0;
             if (isset($_POST['thanhtoantatca'])) {
-                $select_giohang="select * from `giohang`";
-                $select_giohang_run=mysqli_query($con,$select_giohang);
-                while($row=mysqli_fetch_assoc($select_giohang_run))
-                {
-                    $get_id=$row['MASP'];
-                    $get_soluong=$row['SOLUONG'];
+                $select_giohang = "select * from `giohang`";
+                $select_giohang_run = mysqli_query($con, $select_giohang);
+                while ($row = mysqli_fetch_assoc($select_giohang_run)) {
+                    $get_id = $row['MASP'];
+                    $get_soluong = $row['SOLUONG'];
                     $select_sanpham = "select * from `sanpham` where MASP='$get_id'";
-                    
+
                     $select_sanpham_run = mysqli_query($con, $select_sanpham);
                     $row_sanpham = mysqli_fetch_assoc($select_sanpham_run);
                     $get_gia = $row_sanpham['GIA'];
                     $total += $get_soluong * $get_gia;
-                    
                 }
                 $get_username = $_SESSION['username'];
                 // lấy mã khách hàng
@@ -240,41 +239,51 @@ cart();
                 $select_khachhang_run = mysqli_query($con, $select_khachhang);
                 $row_khachhang = mysqli_fetch_assoc($select_khachhang_run);
                 $get_makh = $row_khachhang['MAKH'];
-                //chèn hóa đơn
-                date_default_timezone_set('Asia/Ho_Chi_Minh');
-                $get_date = date('Y:m:d H:i:s');
-                $insert_hoadon = "insert into `hoadon` (NGHD,MAKH,TRIGIA,TRANGTHAI) values ('$get_date','$get_makh','$total','-2')";
-                $insert_hoadon_run = mysqli_query($con, $insert_hoadon);
-                //lấy số hd mới dc tạo
-                $select_hoadon = "select MAX(SOHD) from `hoadon`";
-                $select_hoadon_run = mysqli_query($con, $select_hoadon);
-                $row_hoadon = mysqli_fetch_assoc($select_hoadon_run);
-                $get_sohd = $row_hoadon['MAX(SOHD)'];
-                $select_tesst="select * from `giohang`";
-                $select_tesst_run=mysqli_query($con,$select_tesst);
-                while($row_=mysqli_fetch_assoc($select_tesst_run))
-                {
-                    echo var_dump($row_);
-                    //lấy số lượng
-                    $select_giohang = "select * from `giohang`";
-                    $select_giohang_run = mysqli_query($con, $select_giohang);
-                    $row_giohang = mysqli_fetch_assoc($select_giohang_run);
-                    $get_sl = $row_giohang['SOLUONG'];
-                    $get_id=$row_giohang['MASP'];
-                    //chèn vào cthd
-                    $insert_cthd = "insert into `cthd` (SOHD,MASP,SL) value ('$get_sohd','$get_id','$get_sl')";
-                    
-                    $insert_cthd_run = mysqli_query($con, $insert_cthd);
-                    //chèn xong thì xóa
-                    $delete_query = "delete from `giohang` where MASP='$get_id'";
-                    $delete_query = mysqli_query($con, $delete_query);
-                    if ($delete_query) {
-                        echo "<script>window.open('page_shopping_cart.php','_self') </script>";
+
+                //Kiểm tra số dư
+                $sql_sodu = "select * from `khachhang` where MAKH = '$get_makh'";
+                $run = mysqli_query($con, $sql_sodu);
+                $row_sodu = mysqli_fetch_assoc($run);
+                $get_sodu = $row_sodu['SODU'];
+                if ($get_sodu > $total) {
+                    //chèn hóa đơn
+                    date_default_timezone_set('Asia/Ho_Chi_Minh');
+                    $get_date = date('Y:m:d H:i:s');
+                    $insert_hoadon = "insert into `hoadon` (NGHD,MAKH,TRIGIA,TRANGTHAI) values ('$get_date','$get_makh','$total','-2')";
+                    $insert_hoadon_run = mysqli_query($con, $insert_hoadon);
+                    //lấy số hd mới dc tạo
+                    $select_hoadon = "select MAX(SOHD) from `hoadon`";
+                    $select_hoadon_run = mysqli_query($con, $select_hoadon);
+                    $row_hoadon = mysqli_fetch_assoc($select_hoadon_run);
+                    $get_sohd = $row_hoadon['MAX(SOHD)'];
+                    $select_tesst = "select * from `giohang`";
+                    $select_tesst_run = mysqli_query($con, $select_tesst);
+                    while ($row_ = mysqli_fetch_assoc($select_tesst_run)) {
+                        echo var_dump($row_);
+                        //lấy số lượng
+                        $select_giohang = "select * from `giohang`";
+                        $select_giohang_run = mysqli_query($con, $select_giohang);
+                        $row_giohang = mysqli_fetch_assoc($select_giohang_run);
+                        $get_sl = $row_giohang['SOLUONG'];
+                        $get_id = $row_giohang['MASP'];
+                        //chèn vào cthd
+                        $insert_cthd = "insert into `cthd` (SOHD,MASP,SL) value ('$get_sohd','$get_id','$get_sl')";
+
+                        $insert_cthd_run = mysqli_query($con, $insert_cthd);
+                        //chèn xong thì xóa
+                        $delete_query = "delete from `giohang` where MASP='$get_id'";
+                        $delete_query = mysqli_query($con, $delete_query);
+                        if ($delete_query) {
+                            echo "<script>window.open('page_shopping_cart.php','_self') </script>";
+                        }
                     }
+                } else {
+                    echo "<script>alert('Số dư củ của bạn không đủ! Vui lòng nạp thêm tiền để thanh toán!')</script>";
+                    echo "<script>window.open('user.php?payment') </script>";
                 }
             }
         }
-        echo $thanhtoan=thanhtoantatca();
+        echo $thanhtoan = thanhtoantatca();
         ?>
     </div>
 </div>
